# reduce allocations and make things go faster