--- a/src/Esprima/Ast/Literal.cs
+++ b/src/Esprima/Ast/Literal.cs
@@ -75,7 +75,8 @@ namespace Esprima.Ast
 
         string PropertyKey.GetKey()
         {
-            return Convert.ToString(Value);
+            var s = Value as string;
+            return s ?? Convert.ToString(Value);
         }
     }
 }