--- a/src/Esprima/Ast/Node.cs
+++ b/src/Esprima/Ast/Node.cs
@@ -5,6 +5,8 @@ namespace Esprima.Ast
 {
     public class Node : INode
     {
+        private Location _location;
+
         [JsonProperty(Order = -100)]
         [JsonConverter(typeof(StringEnumConverter))]
         public Nodes Type { get; set; }
@@ -12,6 +14,10 @@ namespace Esprima.Ast
         public int[] Range { get; set; }
 
         [JsonProperty(PropertyName = "Loc")]
-        public Location Location { get; set; } = new Location();
+        public Location Location
+        {
+            get => _location  = _location ?? new Location();
+            set => _location = value;
+        }
     }
 }