… src/Esprima/JavascriptParser.cs
… src/Esprima/ParserExtensions.cs