Based on addition of base benchmark I've created the pull request I've gone through code as a fun excercise what would be ways to make it run faster with less memory usage. It seems I was able to almost halve the memory usage and run duration dropped around 25-30%.

Here's my changes:

* cache Func<T> instances - memory pressure analysis showed that a lot of instances were created as code is looping in a fine level and call methods with Func<T> prameters 
* remove string.Substring / ToString allocations where possible & use common lookup  - target grammar is  known and most used words are easy to find, use shared string for known probable words
* reuse Token for object initializer parsing

I can also break to smaller pull requests if needed.

*Before*

|       Method |            FileName |      Mean |     Error |    StdDev |     Gen 0 |     Gen 1 |    Gen 2 | Allocated |
|------------- |-------------------- |----------:|----------:|----------:|----------:|----------:|---------:|----------:|
| **ParseProgram** |       **angular-1.2.5** | **44.729 ms** | **0.7207 ms** | **0.9621 ms** | **3500.0000** | **1000.0000** | **312.5000** |  **19.29 MB** |
| **ParseProgram** |      **backbone-1.1.0** |  **3.817 ms** | **0.0724 ms** | **0.0743 ms** |  **539.0625** |  **265.6250** |        **-** |   **3.08 MB** |
| **ParseProgram** |        **jquery-1.9.1** | **36.866 ms** | **0.4396 ms** | **0.3671 ms** | **3062.5000** |  **937.5000** | **312.5000** |  **16.71 MB** |
| **ParseProgram** | **jquery.mobile-1.4.2** | **60.237 ms** | **1.1951 ms** | **1.0594 ms** | **4625.0000** | **1437.5000** | **500.0000** |  **25.12 MB** |
| **ParseProgram** |      **mootools-1.4.5** | **28.664 ms** | **0.1293 ms** | **0.1010 ms** | **2562.5000** |  **843.7500** | **250.0000** |  **13.84 MB** |
| **ParseProgram** |    **underscore-1.5.2** |  **3.087 ms** | **0.0634 ms** | **0.0778 ms** |  **500.0000** |  **199.2188** |        **-** |    **2.7 MB** |
| **ParseProgram** |          **yui-3.12.0** | **28.712 ms** | **0.7208 ms** | **2.0911 ms** | **2375.0000** |  **812.5000** | **250.0000** |  **12.75 MB** |

*After*

|       Method |            FileName |      Mean |     Error |    StdDev |     Gen 0 |    Gen 1 |    Gen 2 | Allocated |
|------------- |-------------------- |----------:|----------:|----------:|----------:|---------:|---------:|----------:|
| **ParseProgram** |       **angular-1.2.5** | **34.249 ms** | **0.4522 ms** | **0.4230 ms** | **1937.5000** | **750.0000** | **250.0000** |  **10.16 MB** |
| **ParseProgram** |      **backbone-1.1.0** |  **2.829 ms** | **0.0547 ms** | **0.0585 ms** |  **285.1563** | **140.6250** |        **-** |   **1.56 MB** |
| **ParseProgram** |        **jquery-1.9.1** | **25.294 ms** | **0.5014 ms** | **1.0577 ms** | **1589.5833** | **650.0000** | **122.9167** |   **8.66 MB** |
| **ParseProgram** | **jquery.mobile-1.4.2** | **40.444 ms** | **0.8464 ms** | **0.7917 ms** | **2437.5000** | **937.5000** | **250.0000** |  **13.22 MB** |
| **ParseProgram** |      **mootools-1.4.5** | **21.072 ms** | **0.2207 ms** | **0.2064 ms** | **1312.5000** | **562.5000** | **125.0000** |   **7.14 MB** |
| **ParseProgram** |    **underscore-1.5.2** |  **2.280 ms** | **0.0447 ms** | **0.0549 ms** |  **281.2500** |  **74.2188** |        **-** |   **1.36 MB** |
| **ParseProgram** |          **yui-3.12.0** | **16.449 ms** | **0.3290 ms** | **0.3378 ms** | **1093.7500** | **531.2500** |        **-** |   **6.44 MB** |